Move new protocol test into message_compat_tests
